feat(cockpit): parallelize cold-start reconciler via JoinSet + Semaphore
Lifts the cockpit worker reconciler out of `src/server/mod.rs`
(already 2,200+ lines) into its own `src/server/cockpit_reconciler.rs`
module and replaces the serial `for` loop with a `tokio::task::JoinSet`
gated by a `tokio::sync::Semaphore` cap derived from
`cockpit.max_concurrent_resumes`.

Behavior changes vs the old serial loop:

- Targets that pass the per-tick filter (not in `attempted`, not
  already running) get inserted into `attempted` synchronously,
  then dispatched as parallel tasks; an attach-timeout outcome
  clears the entry so the next tick retries cleanly.
- The semaphore caps concurrent in-flight resumes; the supervisor's
  per-agent install gate (separate commit) protects the
  claude-agent-acp lazy-install race for the first spawn per agent.
- The orphan sweep + restart-pending reaping run unchanged before
  the parallel phase; the ORDERING INVARIANT is preserved (sweep
  must precede capacity-counted spawns).

Expected impact: cold start with 12 persisted cockpit sessions goes
from ~30s sequential (≈2.5s/session) to ~3s parallel under the
default cap of 4. First-ever-after-install cold start adds ~2-3s for
the gate-serialised warm-up spawn; every subsequent restart sees the
cached SDK binary and runs fully parallel from frame 0.

Closes #1088 layer A.
